+++ R/zzz.R
@@ -280,3 +280,34 @@ withCallingHandlers_and_choose_call = function(expr, function_names, name_to_dis
   )
   eval(wch, envir=parent.frame())
 }
+
+# function to determine target class and sample value for union, intersect, setdiff, setequal, min, max, range, sum, prod, c, cbind and rbind functions
+target_class_and_sample_value = function(x, recursive=FALSE, errorClasses="") {
+  
+  getClassesOfElements = function(x, recursive, errorClasses) {
+    classes = unlist(lapply(x, function(el) if (class(el)[1L] == "list" || "data.frame" %in% class(el)) "list" else class(el)[1]))
+    if (recursive) {
+      union(classes[classes != "list"], unlist(lapply(x[classes == "list"], function(el) getClassesOfElements(el, recursive=TRUE, errorClasses=errorClasses))))
+    } else {
+      unique(classes)
+    }
+  }
+  classes = getClassesOfElements(x, recursive=isTRUE(recursive), errorClasses=errorClasses)
+  if (length(sel <- intersect(errorClasses, classes)))
+    stop(errorCondition(sprintf(gettext("invalid 'type' (%s) of argument", domain="R"), sel[1L]), call=sys.call(max(sys.nframe() - 1L, 1L))))
+  
+  if (any(c("character", "factor", "ordered") %in% classes)) {
+    valueClass = "character"
+    funValue = character(1L)
+  } else if ("complex" %in% classes) {
+    valueClass = "complex"
+    funValue = complex(1L)
+  } else if (any(c("Date", "POSIXct", "POSIXlt", "difftime") %in% classes)) {
+    valueClass = "double"
+    funValue = numeric(1L)
+  } else {
+    valueClass = "integer64"
+    funValue = integer64(1L)
+  }
+  list(class = valueClass, sampleValue = funValue)
+}